Split Admin Custom User Status components
