+++ client/admin/customUserStatus/EditCustomUserStatus.js
@@ -1,83 +1,13 @@
 import React, { useCallback, useState, useMemo, useEffect } from 'react';
-import { Box, Button, ButtonGroup, TextInput, Field, Select, Icon, Skeleton, Throbber, InputBox, Modal } from '@rocket.chat/fuselage';
+import { Button, ButtonGroup, TextInput, Field, Select, Icon } from '@rocket.chat/fuselage';
 
 import { useTranslation } from '../../contexts/TranslationContext';
 import { useMethod } from '../../contexts/ServerContext';
 import { useToastMessageDispatch } from '../../contexts/ToastMessagesContext';
 import { useSetModal } from '../../contexts/ModalContext';
-import { useEndpointDataExperimental, ENDPOINT_STATES } from '../../hooks/useEndpointDataExperimental';
 import VerticalBar from '../../components/basic/VerticalBar';
-
-const DeleteWarningModal = ({ onDelete, onCancel, ...props }) => {
-	const t = useTranslation();
-	return <Modal {...props}>
-		<Modal.Header>
-			<Icon color='danger' name='modal-warning' size={20}/>
-			<Modal.Title>{t('Are_you_sure')}</Modal.Title>
-			<Modal.Close onClick={onCancel}/>
-		</Modal.Header>
-		<Modal.Content fontScale='p1'>
-			{t('Custom_User_Status_Delete_Warning')}
-		</Modal.Content>
-		<Modal.Footer>
-			<ButtonGroup align='end'>
-				<Button ghost onClick={onCancel}>{t('Cancel')}</Button>
-				<Button primary danger onClick={onDelete}>{t('Delete')}</Button>
-			</ButtonGroup>
-		</Modal.Footer>
-	</Modal>;
-};
-
-const SuccessModal = ({ onClose, ...props }) => {
-	const t = useTranslation();
-	return <Modal {...props}>
-		<Modal.Header>
-			<Icon color='success' name='checkmark-circled' size={20}/>
-			<Modal.Title>{t('Deleted')}</Modal.Title>
-			<Modal.Close onClick={onClose}/>
-		</Modal.Header>
-		<Modal.Content fontScale='p1'>
-			{t('Custom_User_Status_Has_Been_Deleted')}
-		</Modal.Content>
-		<Modal.Footer>
-			<ButtonGroup align='end'>
-				<Button primary onClick={onClose}>{t('Ok')}</Button>
-			</ButtonGroup>
-		</Modal.Footer>
-	</Modal>;
-};
-
-export function EditCustomUserStatusWithData({ _id, cache, ...props }) {
-	const t = useTranslation();
-	const query = useMemo(() => ({
-		query: JSON.stringify({ _id }),
-		// TODO: remove cache. Is necessary for data invalidation
-	}), [_id, cache]);
-
-	const { data, state, error } = useEndpointDataExperimental('custom-user-status.list', query);
-
-	if (state === ENDPOINT_STATES.LOADING) {
-		return <Box pb='x20'>
-			<Skeleton mbs='x8'/>
-			<InputBox.Skeleton w='full'/>
-			<Skeleton mbs='x8'/>
-			<InputBox.Skeleton w='full'/>
-			<ButtonGroup stretch w='full' mbs='x8'>
-				<Button disabled><Throbber inheritColor/></Button>
-				<Button primary disabled><Throbber inheritColor/></Button>
-			</ButtonGroup>
-			<ButtonGroup stretch w='full' mbs='x8'>
-				<Button primary danger disabled><Throbber inheritColor/></Button>
-			</ButtonGroup>
-		</Box>;
-	}
-
-	if (error || !data || data.statuses.length < 1) {
-		return <Box fontScale='h1' pb='x20'>{t('Custom_User_Status_Error_Invalid_User_Status')}</Box>;
-	}
-
-	return <EditCustomUserStatus data={data.statuses[0]} {...props}/>;
-}
+import DeleteWarningModal from './DeleteWarningModal';
+import SuccessModal from './SuccessModal';
 
 export function EditCustomUserStatus({ close, onChange, data, ...props }) {
 	const t = useTranslation();
@@ -163,3 +93,5 @@ export function EditCustomUserStatus({ close, onChange, data, ...props }) {
 		</Field>
 	</VerticalBar.ScrollableContent>;
 }
+
+export default EditCustomUserStatus;
